swars: Introduced map_coords_limit() to handle off-map coords
The game was behaving strangely when given off-map coordinates.
Now we have a function to properly limit the corrdinates, if we
really need a specific cat of coords to be valid on-map coords.

+++ src/thing.c
@@ -1029,7 +1029,7 @@ short add_static(int x, int y, int z, ushort frame, int timer)
 
 TbBool thing_is_within_circle(ThingIdx thing, short X, short Z, ushort R)
 {
-    long dtX, dtZ, r2;
+    s32 dtX, dtZ, r2;
 
     if (thing <= 0) {
         struct SimpleThing *p_sthing;
